# documentation/api-documentation.yml
openapi: "3.0.0"
info:
  version: "3.0.1"
  title: "FemDevs API Documentation"
  description: |-
    The documentation for the FemDevs API.
    Do note that some routes have permission-based requirements that your user must be quilified for in order to access them.
    Check the description for each route in case you are unsure.

    Permissions Overrides:

    The following permissions will bypass all permissions unless its marked otherwise:
    * `Global::Role.Developer`
    * `Global::Role.Administrator`
    * `Global::Role.Owner`
  termsOfService: "https://legal.thefemdevs.com/terms"
  contact:
    name: "FemDevs"
    url: "https://thefemdevs.com"
    email: "support@thefemdevs.com"
  license:
    name: "GNU Affero General Public License v3.0"
    url: "https://www.gnu.org/licenses/agpl-3.0.en.html"
servers:
  - url: "https://api.thefemdevs.com/"
    description: "Production Server"
paths:
...
  /minecraft/hive/player:
    get:
      operationId: "GetHivePlayer"
      description: |-
        Gets the player information for the given username.
        Returns the player information if the player exists, otherwise returns an error.

        Permissions Required: `Minecraft::Hive.Player`
      summary: "Get Hive Player Info"
      parameters:
        - in: "header"
          name: "x-user"
          description: "The Minecraft username of the player to get information for"
          schema:
            type: string
      responses:
        200:
          description: "The player information for the given username"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MinecraftHivePlayerCombinedData"
              examples:
                user:
                  value:
                    hide:
                      xp: 1612
                      played: 13
                      victories: 3
                      first_played: 1600464717
                      deaths: 9
                      hider_kills: 28
                      seeker_kills: 2
                    main:
                      xuid: 2535441321443031
                      username: therealbenpai
                      username_cc: therealbenpai
                      rank: REGULAR
                      first_played: 1597106900
                      daily_login_streak: 0
                      longest_daily_login_streak: 6
                      hub_title_count: 6
                      hub_title_unlocked:
                        - "&cThe Short Sighted"
                        - "&cThe Defender"
                        - "&eAir &aSwimmer"
                        - "&aThe Good Citizen"
                        - "&aThe Snail"
                        - "&eClickety &6Clack"
                      avatar_count: 6
                      avatar_unlocked:
                        - url: https://cdn.playhive.com/avatars/monster-drew.png
                          name: Drew (Monster)
                        - url: https://cdn.playhive.com/avatars/pet-endolotl.png
                          name: Endolotl Pet
                        - url: https://cdn.playhive.com/avatars/monster-harper.png
                          name: Harper (Monster)
                        - url: https://cdn.playhive.com/avatars/monster-reese.png
                          name: Reese (Monster)
                        - url: https://cdn.playhive.com/avatars/monster-riley.png
                          name: Riley (Monster)
                        - url: https://cdn.playhive.com/avatars/grav-zoomies.png
                          name: Zoomies
                      costume_count: 0
                      friend_count: 8
                      equipped_hub_title: "&eClickety &6Clack"
                      equipped_avatar:
                        url: https://cdn.playhive.com/avatars/grav-zoomies.png
                        name: Zoomies
                      quest_count: 20
                      pets:
                        - Rock
                        - Endolotl
                      mounts: []
                      hats: []
                    dr:
                      xp: 4248
                      played: 46
                      victories: 31
                      first_played: 1694643520
                      deaths: 203
                      checkpoints: 264
                      activated: 76
                      kills: 86
                    wars:
                      xp: 1160
                      played: 27
                      first_played: 1612317414
                      final_kills: 5
                      kills: 27
                      treasure_destroyed: 19
                      deaths: 82
                      victories: 1
                    murder:
                      xp: 16914
                      played: 420
                      victories: 308
                      first_played: 1597106959
                      deaths: 192
                      coins: 3193
                      murders: 170
                      murderer_eliminations: 22
                    sg:
                      xp: 65
                      played: 6
                      first_played: 1615497464
                      crates: 40
                      kills: 6
                      deaths: 6
                      cows: 1
                    sky:
                      xp: 320
                      played: 27
                      victories: 5
                      first_played: 1612392245
                      ores_mined: 551
                      deaths: 20
                      kills: 5
                      mystery_chests_destroyed: 1
                      spells_used: 10
                    ctf:
                      xp: 143
                      played: 3
                      victories: 1
                      first_played: 1694295639
                      assists: 4
                      deaths: 7
                      flags_captured: 0
                      kills: 6
                      flags_returned: 1
                    drop:
                      xp: 65
                      played: 1
                      first_played: 1694288117
                      blocks_destroyed: 506
                      powerups_collected: 0
                      vaults_used: 0
                      deaths: 1
                    ground:
                      xp: 0
                      played: 0
                      victories: 0
                      first_played: 0
                      blocks_destroyed: 0
                      powerups_collected: 0
                      vaults_used: 0
                      deaths: 0
                    build:
                      xp: 10
                      played: 2
                      first_played: 1597107129
                      rating_good_received: 5
                      rating_love_received: 0
                      rating_meh_received: 4
                      rating_okay_received: 1
                      rating_great_received: 0
                    party:
                      xp: 1072
                      played: 35
                      victories: 2
                      first_played: 1687883920
                      powerups_collected: 55
                      rounds_survived: 284
                    bridge:
                      xp: 0
                      first_played: 0
                    grav:
                      xp: 658
                      played: 21
                      victories: 7
                      first_played: 1694283594
                      deaths: 413
                      maps_completed: 57
                      maps_completed_without_dying: 15
        400:
          description: "Missing user information"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorData"
              examples:
                Missing-User:
                  value:
                    code: 8
                    message: "Missing Header Parameter(s)"
                Invalid-User:
                  value:
                    code: 9
                    message: "Invalid Header Parameter(s)"
        401:
          description: "Invalid/Missing API Token"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorData"
              examples:
                Missing-Token:
                  summary: "Missing API Token"
                  value:
                    code: 3
                    message: "No API key provided"
                Invalid-Token:
                  summary: "Invalid API Token"
                  value:
                    code: 4
                    message: "Invalid API key provided"
        403:
          description: "Missing Permissions"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorData"
              examples:
                Missing-Permissions:
                  value:
                    code: 1
                    message: "You are not authorized to access this resource"
                Blacklisted:
                  value:
                    code: 2
                    message: "You are blacklisted from accessing this resource"
        404:
          description: "Error with the given username"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorData"
              examples:
                error:
                  value:
                    code: 22
                    message: "Unable to retreive requested information"
      security:
        - APIToken: []
  /minecraft/hive/maps:
    get:
      operationId: "GetHiveMaps"
      description: |-
        Gets the list of maps for the given game.
        Returns the list of maps if the game exists, otherwise returns an error.

        Permissions Required: `Minecraft::Hive.Map`
      summary: "Get Hive Maps"
      parameters:
        - in: "header"
          name: "x-game"
          description: "The game to get maps for"
          schema:
            type: string
      responses:
        200:
          description: "The list of maps for the given game"
          content:
            application/json:
              schema:
                type: array
                maxItems: 100
                items:
                  type: object
                  $ref: "#/components/schemas/MinecraftHiveMapData"
              examples:
                maps:
                  value:
                    - name: Chroma
                      season: NO_SEASON
                      variant: REGULAR
                      image: https://cdn.playhive.com/maps/mm_chroma.jpg
                    - name: Eso Hotel
                      season: NO_SEASON
                      variant: REGULAR
                      image: https://cdn.playhive.com/maps/mm_esohotel.jpg
                    - name: Manor
                      season: NO_SEASON
                      variant: REGULAR
                      image: https://cdn.playhive.com/maps/mm_manor.jpg
                    - name: Museum
                      season: NO_SEASON
                      variant: REGULAR
                      image: https://cdn.playhive.com/maps/mm_museum.jpg
                    - name: Office Party
                      season: WINTERFEST
                      variant: REGULAR
                      image: https://cdn.playhive.com/maps/mm_office.winterfest.jpg
                    - name: Office
                      season: NO_SEASON
                      variant: REGULAR
                      image: https://cdn.playhive.com/maps/mm_office.jpg
                    - name: Pigment
                      season: NO_SEASON
                      variant: REGULAR
                      image: https://cdn.playhive.com/maps/mm_pigment.jpg
                    - name: Remnants
                      season: NO_SEASON
                      variant: REGULAR
                      image: https://cdn.playhive.com/maps/mm_remnants.jpg
                    - name: Spaceship
                      season: NO_SEASON
                      variant: REGULAR
                      image: https://cdn.playhive.com/maps/mm_spaceship.jpg
                    - name: Ghost House
                      season: HALLOWEEN
                      variant: REGULAR
                      image: https://cdn.playhive.com/maps/mm_toyhouse.halloween.jpg
                    - name: Toy House
                      season: NO_SEASON
                      variant: REGULAR
                      image: https://cdn.playhive.com/maps/mm_toyhouse.jpg
                    - name: Village
                      season: WINTERFEST
                      variant: REGULAR
                      image: https://cdn.playhive.com/maps/mm_village.winterfest.jpg
                    - name: Warehouse
                      season: NO_SEASON
                      variant: REGULAR
                      image: https://cdn.playhive.com/maps/mm_warehouse.jpg
                    - name: Witch School
                      season: HALLOWEEN
                      variant: REGULAR
                      image: https://cdn.playhive.com/maps/mm_witchschool.halloween.jpg
        400:
          description: "Missing game information"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorData"
              examples:
                Missing-Game:
                  value:
                    code: 8
                    message: "Missing Header Parameter(s)"
                Invalid-Game:
                  value:
                    code: 9
                    message: "Invalid Header Parameter(s)"
        401:
          description: "Invalid/Missing API Token"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorData"
              examples:
                Missing-Token:
                  summary: "Missing API Token"
                  value:
                    code: 3
                    message: "No API key provided"
                Invalid-Token:
                  summary: "Invalid API Token"
                  value:
                    code: 4
                    message: "Invalid API key provided"
        403:
          description: "Missing Permissions"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorData"
              examples:
                Missing-Permissions:
                  value:
                    code: 1
                    message: "You are not authorized to access this resource"
                Blacklisted:
                  value:
                    code: 2
                    message: "You are blacklisted from accessing this resource"
        404:
          description: "Error with the given game"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorData"
              examples:
                error:
                  value:
                    code: 22
                    message: "Unable to retreive requested information"
      security:
        - APIToken: []
...
components:
  securitySchemes:
    APIToken:
      type: "http"
      scheme: "bearer"
      bearerFormat: "JWT"
      description: |-
        JSON Web Token Authentication with information containing your license key

        Permissions:
        - `Global::Data.Read` (0:0:0) - **Unused**
        - `Global::Data.Write` (0:0:1) - **Unused**
        - `Global::Data.Create` (0:0:2) - **Unused**
        - `Global::Data.Delete` (0:0:3) - **Unused**
        - `Global::Token.Read` (0:1:0) - Allows reading of token information
        - `Global::Token.ReadAll` (0:1:1) - Allows reading of all token information
        - `Global::Token.Write` (0:1:2) - **Unused**
        - `Global::Token.Create` (0:1:3) - Allows creating of token information
        - `Global::Token.Delete` (0:1:4) - Allows deleting of token 
        - `Global::User.Read` (0:2:0) - Read user information
        - `Global::User.ReadAll` (0:2:1) - **Unused**
        - `Global::User.Write` (0:2:2) - Write/Update user information
        - `Global::User.Create` (0:2:3) - Create a user profile
        - `Global::User.Delete` (0:2:4) - Delete a user profile
        - `Cryptography::Encrypt` (1:0) - Ability to encrypt data using the `/crypto/enc` endpoint
        - `Cryptography::Decrypt` (1:1) - Ability to decrypt data using the `/crypto/dec` endpoint
        - `Dictionary::Definition` (2:0) - Ability to get the definition of a word using the `/dict/def` endpoint
        - `Dictionary::Synonym` (2:1) - Ability to get the synonyms of a word using the `/dict/syn` endpoint
        - `Dictionary::Antonym` (2:2) - Ability to get the antonyms of a word using the `/dict/ant` endpoint
        - `Facts::Dog` (3:0) - Ability to get a random dog fact using the `/facts/dog` endpoint
        - `Facts::Cat` (3:1) - Ability to get a random cat fact using the `/facts/cat` endpoint
        - `Facts::Random` (3:2) - Ability to get a random fact using the `/facts/random` endpoint
        - `Location::Coord` (4:0) - Ability to get the location of coordinates using the `/location/coords` endpoint
        - `Location::Pluscode` (4:1) - Ability to get the location of a pluscode using the `/location/pluscode` endpoint
        - `Location::Address` (4:2) - Ability to get the location of an address using the `/location/address` endpoint
        - `Phone::Lookup` (5:0) - Ability to get the details of a phone number using the `/phone/details` endpoint
        - `Phone::Validate` (5:1) - **Unused**
        - `Quote::Kanye` (6:0) - Ability to get a random Kanye West quote using the `/quotes/kanye` endpoint
        - `Quote::RonSwanson` (6:1) - Ability to get a random Ron Swanson quote using the `/quotes/ronswanson` endpoint
        - `Quote::Random` (6:2) - Ability to get a random quote using the `/quotes/random` endpoint
        - `Weather::Current` (7:0) - Ability to get the current weather using the `/weather/current` endpoint
        - `Minecraft::Hive.Player` (8:0:0) - Ability to get information about a player on the Hive Minecraft server using the `/minecraft/hive/player` endpoint
...
  schemas:
...
    ErrorData:
      description: |-
        The error data that is returned when an error occurs

        All Error Codes are located in `/functions/errorResponse.js` [GitHub](https://github.com/femdevs/femdev-website/blob/main/functions/errorResponse.js)
      required:
        - code
        - message
      properties:
        code:
          nullable: false
          type: integer
          description: "The error code"
        message:
          nullable: false
          type: string
          description: "The error message"
...
    MinecraftHiveMainPlayerData:
      description: |-
        The Main Player Data for the Hive Minecraft Server
      required:
        - xuid
        - username
        - username_cc
        - rank
        - first_played
        - daily_login_streak
        - longest_daily_login_streak
        - hub_title_count
        - hub_title_unlocked
        - avatar_count
        - avatar_unlocked
        - costume_count
        - friend_count
        - equipped_hub_title
        - equipped_avatar
        - quest_count
        - pets
        - mounts
        - hats
      properties:
        xuid:
          nullable: false
          type: integer
          description: "The XUID of the player"
        username:
          nullable: false
          type: string
          description: "The username of the player"
        username_cc:
          nullable: false
          type: string
          description: "The username of the player with color codes"
        rank:
          nullable: false
          type: string
          description: "The rank of the player"
        first_played:
          nullable: false
          type: integer
          description: "The first time the player played"
        daily_login_streak:
          nullable: false
          type: integer
          description: "The current daily login streak of the player"
        longest_daily_login_streak:
          nullable: false
          type: integer
          description: "The longest daily login streak of the player"
        hub_title_count:
          nullable: false
          type: integer
          description: "The number of hub titles the player has"
        hub_title_unlocked:
          nullable: false
          type: array
          items:
            type: string
          description: "The number of hub titles the player has unlocked"
        avatar_count:
          nullable: false
          type: integer
          description: "The number of avatars the player has"
        avatar_unlocked:
          nullable: false
          type: array
          items:
            type: object
            $ref: "#/components/schemas/MinecraftHiveURLDataObj"
          description: "The avatars the player has unlocked"
        costume_count:
          nullable: false
          type: integer
          description: "The number of costumes the player has"
        friend_count:
          nullable: false
          type: integer
          description: "The number of friends the player has"
        equipped_hub_title:
          nullable: false
          type: string
          description: "The hub title the player has equipped"
        equipped_avatar:
          nullable: false
          type: object
          $ref: "#/components/schemas/MinecraftHiveURLDataObj"
          description: "The avatar the player has equipped"
        quest_count:
          nullable: false
          type: integer
          description: "The number of quests the player has"
        pets:
          nullable: false
          type: array
          items:
            type: string
          description: "The pets the player has"
        mounts:
          nullable: false
          type: array
          items:
            type: string
          description: "The mounts the player has"
        hats:
          nullable: false
          type: array
          items:
            type: string
          description: "The hats the player has"
    MinecraftHivePlayerHideGameStatistics:
      description: |-
        The Hide Game Statistics for the Hive Minecraft Server
      required:
        - xp
        - played
        - victories
        - first_played
        - deaths
        - hider_kills
      properties:
        xp:
          nullable: false
          type: integer
          description: "The XP of the player"
        played:
          nullable: false
          type: integer
          description: "The number of games the player has played"
        victories:
          nullable: false
          type: integer
          description: "The number of games the player has won"
        first_played:
          nullable: false
          type: integer
          description: "The first time the player played"
        deaths:
          nullable: false
          type: integer
          description: "The number of times the player has died"
        hider_kills:
          nullable: false
          type: integer
          description: "The number of hiders the player has killed"
    MinecraftHivePlayerDeathrunStatistics:
      description: |-
        The Deathrun Statistics for the Hive Minecraft Server
      required:
        - xp
        - played
        - victories
        - first_played
        - deaths
        - checkpoints
        - activated
        - kills
      properties:
        xp:
          nullable: false
          type: integer
          description: "The XP of the player"
        played:
          nullable: false
          type: integer
          description: "The number of games the player has played"
        victories:
          nullable: false
          type: integer
          description: "The number of games the player has won"
        first_played:
          nullable: false
          type: integer
          description: "The first time the player played"
        deaths:
          nullable: false
          type: integer
          description: "The number of times the player has died"
        checkpoints:
          nullable: false
          type: integer
          description: "The number of checkpoints the player has reached"
        activated:
          nullable: false
          type: integer
          description: "The number of traps the player has activated"
        kills:
          nullable: false
          type: integer
          description: "The number of deaths the player has caused"
    MinecraftHivePlayerWarStatistics:
      description: |-
        The War Statistics for the Hive Minecraft Server
      required:
        - xp
        - played
        - first_played
        - final_kills
        - kills
        - treasure_destroyed
        - deaths
      properties:
        xp:
          nullable: false
          type: integer
          description: "The XP of the player"
        played:
          nullable: false
          type: integer
          description: "The number of games the player has played"
        first_played:
          nullable: false
          type: integer
          description: "The first time the player played"
        final_kills:
          nullable: false
          type: integer
          description: "The number of final kills the player has"
        kills:
          nullable: false
          type: integer
          description: "The number of kills the player has"
        treasure_destroyed:
          nullable: false
          type: integer
          description: "The number of treasures the player has destroyed"
        deaths:
          nullable: false
          type: integer
          description: "The number of times the player has died"
    MinecraftHivePlayerMurderMysteryStatistics:
      description: |-
        The Murder Mystery for the Minecraft Hive Player
      required:
        - xp
        - played
        - victories
        - first_played
        - deaths
        - coins
        - murders
        - murderer_eliminations
      properties:
        xp:
          nullable: false
          type: integer
          description: "The XP of the player"
        played:
          nullable: false
          type: integer
          description: "The number of games the player has played"
        victories:
          nullable: false
          type: integer
          description: "The number of games the player has won"
        first_played:
          nullable: false
          type: integer
          description: "The first time the player played"
        deaths:
          nullable: false
          type: integer
          description: "The number of times the player has died"
        coins:
          nullable: false
          type: integer
          description: "The number of coins the player has"
        murders:
          nullable: false
          type: integer
          description: "The number of murders the player has committed"
        murderer_eliminations:
          nullable: false
          type: integer
          description: "The number of murderers the user has eliminated"
...
    MinecraftHivePlayerCombinedData:
      required:
        - hide
        - main
        - deathrun
        - war
        - murder
        - sg
        - sky
        - ctf
        - drop
        - ground
        - build
        - party
        - bridge
        - grav
      properties:
        hide:
          nullable: false
          type: object
          $ref: "#/components/schemas/MinecraftHivePlayerHideGameStatistics"
          description: "The hide game statistics of the player"
        main:
          nullable: false
          type: object
          $ref: "#/components/schemas/MinecraftHiveMainPlayerData"
          description: "The main statistics of the player"
        deathrun:
          nullable: false
          type: object
          $ref: "#/components/schemas/MinecraftHivePlayerDeathrunStatistics"
          description: "The deathrun statistics of the player"
        war:
          nullable: false
          type: object
          $ref: "#/components/schemas/MinecraftHivePlayerWarStatistics"
          description: "The war statistics of the player"
        murder:
          nullable: false
          type: object
          $ref: "#/components/schemas/MinecraftHivePlayerMurderMysteryStatistics"
          description: "The murder mystery game statistics of the player"
        sg:
          nullable: false
          type: object
          $ref: "#/components/schemas/MinecraftHivePlayerSurvivalGamesStatistics"
          description: "The survival games statistics of the player"
        sky:
          nullable: false
          type: object
          $ref: "#/components/schemas/MinecraftHivePlayerSkyWarsStatistics"
          description: "The sky wars statistics of the player"
        ctf:
          nullable: false
          type: object
          $ref: "#/components/schemas/MinecraftHivePlayerCTFStatistics"
          description: "The capture the flag statistics of the player"
        drop:
          nullable: false
          type: object
          $ref: "#/components/schemas/MinecraftHivePlayerDropStatistics"
          description: "The drop statistics of the player"
        ground:
          nullable: false
          type: object
          $ref: "#/components/schemas/MinecraftHivePlayerGroundStatistics"
          description: "The ground statistics of the player"
        build:
          nullable: false
          type: object
          $ref: "#/components/schemas/MinecraftHivePlayerBuildStatistics"
          description: "The build statistics of the player"
        party:
          nullable: false
          type: object
          $ref: "#/components/schemas/MinecraftHivePlayerPartyStatistics"
          description: "The party statistics of the player"
        bridge:
          nullable: false
          type: object
          $ref: "#/components/schemas/MinecraftHivePlayerBridgeStatistics"
          description: "The bridge statistics of the player"
        grav:
          nullable: false
          type: object
          $ref: "#/components/schemas/MinecraftHivePlayerGravityStatistics"
          description: "The gravity statistics of the player"
    MinecraftHiveMapData:
      description: |-
        The Map Data for the Hive Minecraft Server
      required:
        - name
        - season
        - variant
        - image
      properties:
        name:
          nullable: false
          type: string
          description: "The name of the map"
        season:
          nullable: false
          type: string
          description: "The season of the map"
        variant:
          nullable: false
          type: string
          description: "The variant of the map"
        image:
          nullable: false
          type: string
          description: "The image of the map"
...